v1/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.0
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
googleapis-common-protos==1.65.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.1
openai==1.43.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-openai==0.28.2
opentelemetry-instrumentation-system-metrics==0.47b0
opentelemetry-propagator-b3==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.1
protobuf==4.25.5
psutil==6.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
splunk-opentelemetry==1.20.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
